Move `usePluginEntities` into `hooks` directory.
