+++ graylog2-web-interface/src/components/common/PublicNotifications.test.tsx
@@ -20,7 +20,7 @@ import type { PluginExports } from 'graylog-web-plugin/plugin';
 
 import type { Notifications } from 'theme/types';
 import { asMock } from 'helpers/mocking';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 
 import PublicNotifications from './PublicNotifications';
 
@@ -73,7 +73,7 @@ const mockedConfigNotifications = {
   },
 } as Notifications;
 
-jest.mock('views/logic/usePluginEntities');
+jest.mock('hooks/usePluginEntities');
 
 jest.mock('util/AppConfig', () => ({
   publicNotifications: jest.fn(() => mockedConfigNotifications),
