+++ graylog2-web-interface/src/contexts/GlobalContextProviders.test.tsx
@@ -20,9 +20,9 @@ import { render, screen } from 'wrappedTestingLibrary';
 import asMock from 'helpers/mocking/AsMock';
 import suppressConsole from 'helpers/suppressConsole';
 import GlobalContextProviders from 'contexts/GlobalContextProviders';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 
-jest.mock('views/logic/usePluginEntities');
+jest.mock('hooks/usePluginEntities');
 
 const renderProvider = (children, index, throwError = false) => {
   if (throwError) {
