+++ graylog2-web-interface/src/hooks/usePluginEntities.ts
