+++ graylog2-web-interface/src/pages/ConfigurationsPage.test.tsx
@@ -22,10 +22,10 @@ import suppressConsole from 'helpers/suppressConsole';
 import MockStore from 'helpers/mocking/StoreMock';
 import { configuration as mockConfiguration } from 'fixtures/configurations';
 import ConfigurationsPage from 'pages/ConfigurationsPage';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 import SidecarConfig from 'components/configurations/SidecarConfig';
 
-jest.mock('views/logic/usePluginEntities');
+jest.mock('hooks/usePluginEntities');
 jest.mock('components/configurations/SearchesConfig', () => () => <span>Search Configuration</span>);
 jest.mock('components/configurations/MessageProcessorsConfig', () => () => <span>Message Processors Configuration</span>);
 jest.mock('components/configurations/SidecarConfig');
