+++ graylog2-web-interface/src/pages/ConfigurationsPage.tsx
@@ -29,7 +29,7 @@ import UrlWhiteListConfig from 'components/configurations/UrlWhiteListConfig';
 import PermissionsConfig from 'components/configurations/PermissionsConfig';
 import 'components/maps/configurations';
 import type { Store } from 'stores/StoreTypes';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 import ConfigletRow from 'pages/configurations/ConfigletRow';
 import { ConfigurationsActions, ConfigurationsStore } from 'stores/configurations/ConfigurationsStore';
 import { CurrentUserStore } from 'stores/users/CurrentUserStore';
