+++ graylog2-web-interface/src/views/components/WidgetQueryControls.tsx
@@ -50,7 +50,7 @@ import {
   pluggableValidationPayload,
 } from 'views/logic/searchbar/pluggableSearchBarControlsHandler';
 import type { SearchBarControl } from 'views/types';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 import useUserDateTime from 'hooks/useUserDateTime';
 
 import TimeRangeOverrideInfo from './searchbar/WidgetTimeRangeOverride';
