+++ graylog2-web-interface/src/views/components/aggregationwizard/visualization/VisualizationConfiguration.tsx
@@ -21,7 +21,7 @@ import styled from 'styled-components';
 
 import { Input, Checkbox } from 'components/bootstrap';
 import Select from 'components/common/Select';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 import { defaultCompare } from 'logic/DefaultCompare';
 import type { WidgetConfigFormValues } from 'views/components/aggregationwizard/WidgetConfigForm';
 import { TIMESTAMP_FIELD } from 'views/Constants';
