+++ graylog2-web-interface/src/views/components/export/ExportFormatSelection.tsx
@@ -17,7 +17,7 @@
 import * as React from 'react';
 import { Field } from 'formik';
 
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 import Select from 'components/common/Select';
 import { defaultCompare } from 'logic/DefaultCompare';
