+++ graylog2-web-interface/src/views/components/messagelist/MessageAugmentations.test.tsx
@@ -19,12 +19,12 @@ import { render, screen } from 'wrappedTestingLibrary';
 
 import asMock from 'helpers/mocking/AsMock';
 import { simpleMessage as message } from 'fixtures/messages';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 import WindowDimensionsContext from 'contexts/WindowDimensionsContext';
 
 import MessageAugmentations from './MessageAugmentations';
 
-jest.mock('views/logic/usePluginEntities');
+jest.mock('hooks/usePluginEntities');
 
 describe('MessageAugmentations', () => {
   const SimpleMessageAugmentation = () => (
