+++ graylog2-web-interface/src/views/components/messagelist/MessagePreview.tsx
@@ -20,7 +20,7 @@ import styled, { css } from 'styled-components';
 import type MessagesWidgetConfig from 'views/logic/widgets/MessagesWidgetConfig';
 import type FieldType from 'views/logic/fieldtypes/FieldType';
 import type { Message } from 'views/components/messagelist/Types';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 import MessageFieldRow from 'views/components/messagelist/MessageFieldRow';
 
 const TableRow = styled.tr(({ theme }) => css`
