+++ graylog2-web-interface/src/views/components/messagelist/MessageTableProviders.test.tsx
@@ -19,11 +19,11 @@ import { render, screen } from 'wrappedTestingLibrary';
 
 import asMock from 'helpers/mocking/AsMock';
 import suppressConsole from 'helpers/suppressConsole';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 
 import MessageTableProviders from './MessageTableProviders';
 
-jest.mock('views/logic/usePluginEntities');
+jest.mock('hooks/usePluginEntities');
 
 const renderProvider = (children, index, throwError = false) => {
   if (throwError) {
