+++ graylog2-web-interface/src/views/components/searchbar/queryinput/QueryInput.tsx
@@ -27,7 +27,7 @@ import type { QueryValidationState } from 'views/components/searchbar/queryvalid
 import useFieldTypes from 'views/logic/fieldtypes/useFieldTypes';
 import { DEFAULT_TIMERANGE } from 'views/Constants';
 import { isNoTimeRangeOverride } from 'views/typeGuards/timeRange';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 import useUserDateTime from 'hooks/useUserDateTime';
 
 import type { AutoCompleter, Editor } from './ace-types';
