+++ graylog2-web-interface/src/views/components/searchbar/queryvalidation/QueryValidation.tsx
@@ -30,7 +30,7 @@ import DocsHelper from 'util/DocsHelper';
 import QueryValidationActions from 'views/actions/QueryValidationActions';
 import FormWarningsContext from 'contexts/FormWarningsContext';
 import type { QueryValidationState } from 'views/components/searchbar/queryvalidation/types';
-import usePluginEntities from 'views/logic/usePluginEntities';
+import usePluginEntities from 'hooks/usePluginEntities';
 
 const Container = styled.div`
   margin-right: 5px;
